pyproject.toml
[project]
name = "sentiment-analysis-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.15.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "praw>=7.7.0",
    "tweepy>=4.14.0",
    "google-api-python-client>=2.184.0",
    "emoji>=2.15.0",
    "nltk>=3.9.2",
    "scikit-learn>=1.7.2",
]
